% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_table.R
\name{basic_table}
\alias{basic_table}
\title{Basic Table
From a list of provided sequencing summaries it generates a database with basic table (contains data such as: run id, duration, number of reads) for each sequencing summary}
\usage{
basic_table(my_data)
}
\arguments{
\item{my_data}{}
}
\description{
Basic Table
From a list of provided sequencing summaries it generates a database with basic table (contains data such as: run id, duration, number of reads) for each sequencing summary
}
