% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n50.R
\name{calculate_n50}
\alias{calculate_n50}
\title{Calculate N50}
\usage{
calculate_n50(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
Numeric N50 value in bases
}
\description{
Calculates the N50 value: the read length at which 50% of all sequenced
bases are contained in reads of that length or longer.
}
\examples{
\dontrun{
 seq_summary <- data.frame(
  sequence_length_template = c(1, 2, 3, 4, 10)
)
calculate_n50(seq_summary)
}
}
