% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n50.R
\name{n50_calc}
\alias{n50_calc}
\title{Calculate N50}
\usage{
n50_calc(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
n50 value
}
\description{
Calculate N50
}
\examples{
\dontrun {
 n50 <- lapply(seq_sum_database, calculate_n50)
}
}
