% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoQuRe-package.R
\docType{package}
\name{NanoQuRe-package}
\alias{NanoQuRe}
\alias{NanoQuRe-package}
\title{NanoQuRe: Quality Control for Oxford Nanopore Sequencing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides quality control tools for Oxford Nanopore Technologies (ONT) sequencing data. Reads sequencing summary files produced by Guppy or Dorado and generates interactive plots and summary tables covering read-length distributions, base quality scores, cumulative yield, sequencing throughput, translocation speed, and pore channel activity. A single call to generate_report() renders a self-contained HTML report from one or more summary files, with automatic single-sample or multi-sample layout selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://LRB-IIMCB.github.io/NanoQuRe}
  \item \url{https://github.com/LRB-IIMCB/NanoQuRe}
  \item Report bugs at \url{https://github.com/LRB-IIMCB/NanoQuRe/issues}
}

}
\author{
\strong{Maintainer}: Natalia Gumińska \email{nguminska@iimcb.gov.pl}

Authors:
\itemize{
  \item Alicja Stachurska \email{ala.stachurska04@gmail.com}
}

}
\keyword{internal}
