% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_yield.R
\name{plot_cumulative_yield}
\alias{plot_cumulative_yield}
\title{Plot Cumulative Yield}
\usage{
<<<<<<< HEAD
plot_cumulative_yield(seq_summary, max_points = 2000L)
=======
plot_cumulative_yield(seq_summary)
>>>>>>> 18b05ca60c14a9aac5e10462f20025840a3c0b08
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
plotly object
}
\description{
Generates an interactive cumulative plot containing the number of sequenced
bases in Gb over time in hours sorted by pass/fail filtering status.
}
\examples{
plot_cumulative_yield(sample_data)
}
