% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_distribution.R
\name{plot_quality_distribution}
\alias{plot_quality_distribution}
\title{Plot Quality Distribution}
\usage{
plot_quality_distribution(seq_summary, qscore_cutoff = 7)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}

\item{qscore_cutoff}{Numeric parameter of Qscore cut-off}
}
\value{
plotly object
}
\description{
Generates an interactive plot with number of reads at each Q score sorted
by pass/fail filtering status. Vertical line represents Q score cut-off
which is by default equal to 7. The basecaller threshold line is only drawn
when both passing and failing reads are present in the data.
}
\examples{
\dontrun{
plot_quality_distribution(sample_data)
}
}
