% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_lengths.R
\name{plot_read_lengths}
\alias{plot_read_lengths}
\title{Plot Read Lengths}
\usage{
plot_read_lengths(seq_summary, upper_limit = 4000, y_limit = NULL)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}

\item{upper_limit}{Upper limit of the x axis. Defaults to 4000.}

\item{y_limit}{Upper limit of the y axis. If NULL (default), automatically
set to the 99.9th percentile of read counts multiplied by 1.1.}
}
\value{
plotly object
}
\description{
Generates an interactive plot with number of reads of each length in bases.
Two vertical lines represent the mean read length and the N50 value
calculated with \code{\link{calculate_n50}}.
}
\examples{
\dontrun{
plot_read_lengths(sample_data)
}
}
