% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_table.R
\docType{data}
\name{plotlytab_database}
\alias{plotlytab_database}
\title{Basic Table
From a list of provided sequencing summaries it generates a database with basic table (contains data such as: run id, duration, number of reads) for each sequencing summary}
\format{
An object of class \code{list} of length 0.
}
\usage{
plotlytab_database
}
\arguments{
\item{my_data}{}
}
\description{
Basic Table
From a list of provided sequencing summaries it generates a database with basic table (contains data such as: run id, duration, number of reads) for each sequencing summary
}
\keyword{datasets}
