% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_distribution.R
\name{quality_distr}
\alias{quality_distr}
\title{Plot Quality Distribution}
\usage{
quality_distr(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
plot
}
\description{
Plot Quality Distribution
}
\examples{
\dontrun {
 plot_quality <- lapply(seq_sum_database, plot_quality_distribution)
}

}
