% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_table.R
\name{quality_table}
\alias{quality_table}
\title{Quality table
collects and summarizes most important quality metrics from sequencing summary files including: sample id, all reads, passed reads, n50 value, mean qscore, longest read, passed mean length}
\usage{
quality_table(my_data)
}
\arguments{
\item{my_data}{}
}
\description{
Quality table
collects and summarizes most important quality metrics from sequencing summary files including: sample id, all reads, passed reads, n50 value, mean qscore, longest read, passed mean length
}
