% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_stats.R
\name{seq_stats}
\alias{seq_stats}
\title{Sequencing Stats}
\usage{
seq_stats(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
return joined table with rudimentary data for all sequencing summaries in database
}
\description{
From sequencing summary's database creates a joined table containing general metrics regarding each run - each sequencing summary.
}
\examples{
\dontrun {
 tab1 <- lapply(seq_sum_database, basic_table)
}
}
