% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorter_seqsum.R
\name{shorter_seqsum}
\alias{shorter_seqsum}
\title{Shorter Seq sum}
\usage{
shorter_seqsum(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
Sequencing summary data frame
}
\description{
Function takes user's sequencing summary and makes a shorter version of sequencing_summary.txt
}
\examples{
NULL
}
