% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_table.R
\name{basic_table}
\alias{basic_table}
\title{Basic Table
From sequencing summary's database creates a joined table containing general metrics regarding each run - each sequencing summary.}
\usage{
basic_table(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
return joined table with rudimentary data for all sequencing summaries in database
}
\description{
Basic Table
From sequencing summary's database creates a joined table containing general metrics regarding each run - each sequencing summary.
}
\examples{
\dontrun {
 tab1 <- lapply(seq_sum_database, basic_table)
}
}
