% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n50.R
\name{calculate_n50}
\alias{calculate_n50}
\title{Calculate N50
Calculates N50 value is the length which 50% of all reads are the same length or longer than its value}
\usage{
calculate_n50(my_data)
}
\arguments{
\item{my_data}{A dataframe containing the sequencing summary}
}
\value{
n50 value
}
\description{
Calculate N50
Calculates N50 value is the length which 50% of all reads are the same length or longer than its value
}
