% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_n50.R
\name{calculate_n50}
\alias{calculate_n50}
\title{Calculate N50}
\usage{
calculate_n50(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
n50 value
}
\description{
Calculates N50 value is the length which 50% of all reads are the same length or longer than its value
}
\examples{
NULL
}
