% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Nanopore QC Report}
\usage{
generate_report(my_data, output_file = "QC_Report.html")
}
\arguments{
\item{my_data}{seq summary file}

\item{output_file}{generated file html}
}
\description{
Generate Nanopore QC Report
}
