% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_active_channels.R
\name{plot_active_channels}
\alias{plot_active_channels}
\title{Plot Active Channels}
\usage{
plot_active_channels(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
plotly object
}
\description{
Generates an interactive plot showing number of active channels during
sequencing time.
}
\examples{
NULL
}
