% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_speed.R
\name{plot_average_speed}
\alias{plot_average_speed}
\title{Plot Average Speed}
\usage{
plot_average_speed(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
plotly object
}
\description{
Generates an interactive plot with average speed of bases going through
the pore per hour, filtered by pass/fail status.
}
\examples{
NULL
}
