% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_speed.R
\name{plot_average_speed}
\alias{plot_average_speed}
\title{Plot Average Speed}
\usage{
plot_average_speed(my_data)
}
\arguments{
\item{my_data}{A dataframe containing the sequencing summary}
}
\value{
ggplot2 object
}
\description{
Generates a plot with average speed of bases going through the pore, filtered by pass status.
}
\examples{
NULL
}
