% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_yield.R
\name{plot_cumulative_yield}
\alias{plot_cumulative_yield}
\title{Plot Cumulative Yield}
\usage{
plot_cumulative_yield(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
ggplot2 object
}
\description{
Generates a cumulative plot containing the number of sequenced bases in Gb over time in hours sorted by pass/fail filtering status.
}
\examples{
\dontrun {
 plot_yield <- lapply(seq_sum_database, plot_cumulative_yield)
}
}
