% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_yield.R
\name{plot_cumulative_yield}
\alias{plot_cumulative_yield}
\title{Plot Cumulative Yield}
\usage{
plot_cumulative_yield(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
ggplot2 object

plotly object
}
\description{
Generates a cumulative plot containing the number of sequenced bases in Gb over time in hours sorted by pass/fail filtering status.
}
\examples{
NULL
Plot Cumulative Yield

Generates an interactive cumulative plot containing the number of sequenced
bases in Gb over time in hours sorted by pass/fail filtering status.

NULL
}
