% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_double.R
\name{plot_double}
\alias{plot_double}
\title{Plot Double}
\usage{
plot_double(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
plotly subplot object
}
\description{
Generates an interactive combined plot containing minimum, maximum and
average read length over sequencing time (log10 scale) and minimum, maximum
and average Q score over sequencing time, displayed side by side.
}
\examples{
NULL
}
