% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_double.R
\name{plot_double}
\alias{plot_double}
\title{Plot Double}
\usage{
plot_double(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
ggplot2 object
}
\description{
Generates a plot containing minimum, maximum and average read length over sequencing time and a plot containing minimum, maximum and average value of qscore through sequencing.
}
\examples{
NULL
}
