% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_lengths.R
\name{plot_read_lengths}
\alias{plot_read_lengths}
\title{Plot Read Lengths}
\usage{
plot_read_lengths(seq_summary, upper_limit = 4000, y_limit = NULL)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}

\item{upper_limit}{limit of x axis, default option is 4000}
}
\value{
ggplot2 object
}
\description{
Generates a plot with number of bases sequenced of each length in bases sorted by pass/fail filtering status.Two vertical lines represent calculated mean read length of the reads in bases and N 50 value calculated with the function calculate_n50.
}
\examples{
NULL
}
