% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_lengths.R
\name{plot_read_lengths}
\alias{plot_read_lengths}
\title{Plot Read Lengths}
\usage{
plot_read_lengths(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
plot
}
\description{
Plot Read Lengths
}
\examples{
\dontrun {
 plot_length <- lapply(seq_sum_database, plot_read_lengths)
}
}
