% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seq_throughput.R
\name{plot_seq_throughput}
\alias{plot_seq_throughput}
\title{Plot Seq Throughput}
\usage{
plot_seq_throughput(my_data)
}
\arguments{
\item{my_data}{A dataframe containing the sequencing summary}
}
\value{
ggplot2 object
}
\description{
Generates a plot with sequence throughput during time filtered by pass status
}
\examples{
NULL
}
