% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seq_throughput.R
\name{plot_seq_throughput}
\alias{plot_seq_throughput}
\title{Plot Seq Throughput}
\usage{
plot_seq_throughput(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
plotly object
}
\description{
Generates an interactive plot with sequence throughput during time
filtered by pass/fail status.
}
\examples{
NULL
}
