% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pore_activity_heatmap.R
\name{pore_activity_heatmap}
\alias{pore_activity_heatmap}
\title{Pore Activity Heatmap}
\usage{
pore_activity_heatmap(seq_summary, platform = "minion")
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}

\item{platform}{Flowcell used for sequencing ("minion" or "promethion"). Defaults to "minion"}
}
\value{
plotly object
}
\description{
Generates an interactive heatmap showing channel activity over time and
channel activity with respect to their disposition on the flowcell.
}
\examples{
NULL
}
