% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quality_distribution.R
\name{quality_distr}
\alias{quality_distr}
\title{Plot Quality Distribution
Generates a plot with number of reads of each length and their average quality score Q score sorted by pass/fail filtering status. Vertical line represents Q score cut-off which is by default equal to 7.}
\usage{
quality_distr(my_data)
}
\arguments{
\item{my_data}{A dataframe containing the sequencing summary}
}
\value{
ggplot2 object
}
\description{
Plot Quality Distribution
Generates a plot with number of reads of each length and their average quality score Q score sorted by pass/fail filtering status. Vertical line represents Q score cut-off which is by default equal to 7.
}
