% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_stats.R
\name{quality_stats}
\alias{quality_stats}
\title{Quality Stats}
\usage{
quality_stats(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
A dataframe with quality metrics for the sequencing run
}
\description{
Generates a table containing the most important quality metrics from a
sequencing summary: number of all reads, number of passed reads, N50 value,
mean Q-score, longest read length, and mean length of passed reads.
}
\examples{
quality_stats(sample_data)
}
