% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_stats.R
\name{quality_stats}
\alias{quality_stats}
\title{Quality Stats
Generates a joined table from sequencing summaries that contains most import quality metrics - number of all reads, number of passed reads, n50 values, mean q-score, longest read length, mean length of passed reads.}
\usage{
quality_stats(my_data)
}
\arguments{
\item{my_data}{}
}
\value{
joined table with quality data for all sequencing summaries in database
}
\description{
Quality Stats
Generates a joined table from sequencing summaries that contains most import quality metrics - number of all reads, number of passed reads, n50 values, mean q-score, longest read length, mean length of passed reads.
}
\examples{
\dontrun {
 tab2 <- lapply(seq_sum_database, quality_table)
}

}
