% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_stats.R
\name{sequencing_stats}
\alias{sequencing_stats}
\title{Sequencing Stats}
\usage{
sequencing_stats(seq_summary)
}
\arguments{
\item{seq_summary}{A dataframe containing the sequencing summary}
}
\value{
A dataframe with general run metrics
}
\description{
Generates a table containing general run metrics from a sequencing summary:
sample ID, duration in hours, number of reads, total bases sequenced in Gb,
percentage of passed reads, and average sequencing speed in bp/s.
}
\examples{
sequencing_stats(sample_data)
}
