% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_stats.R
\name{sequencing_stats}
\alias{sequencing_stats}
\title{Sequencing Stats}
\usage{
sequencing_stats(my_data)
}
\arguments{
\item{my_data}{A dataframe containing the sequencing summary}
}
\value{
returns joined table with rudimentary data for all sequencing summaries in database
}
\description{
From sequencing summary's database creates a joined table containing general metrics regarding each run - each sequencing summary. Information in the table contains: sample id, run id, duration in hours, number of reads, total bases sequenced in GB gigabases and passed_reads in percentage.
}
\examples{
NULL
}
